+++ Hotel/Hotel/Models/GuestServiceUsage.cs
@@ -7,6 +7,6 @@ public class GuestServiceUsage
     public int ServiceId { get; set; }
     public int ReservationId { get; set; }
     public int Quantity { get; set; }
-    public float Price { get; set; }
+    public Double Price { get; set; }
     public DateTime UseDate { get; set; }
 }
