+++ Hotel/Hotel/Models/MenuItem.cs
@@ -6,5 +6,5 @@ public class MenuItem
     public int MenuCategoryId { get; set; }
     public string? Name { get; set; }
     public string? Description { get; set; }
-    public float Price { get; set; }
+    public Double Price { get; set; }
 }
