+++ Hotel/Hotel/Models/Payment.cs
@@ -5,7 +5,7 @@ public class Payment
     public int Id { get; set; }
     public int? InvoiceId { get; set; }
     public int? OrderId { get; set; }
-    public float Amount { get; set; }
+    public Double Amount { get; set; }
     public string? Status { get; set; }
     public DateTime PaymentDate { get; set; }
     public string? TransactionId { get; set; }
