+++ Hotel/Hotel/Models/Position.cs
@@ -4,5 +4,5 @@ public class Position
 {
     public int Id { get; set; }
     public string? Title { get; set; }
-    public float BaseSalary { get; set; }
+    public Double BaseSalary { get; set; }
 }
