+++ Hotel/Hotel/Models/Service.cs
@@ -5,5 +5,5 @@ public class Service
     public int Id { get; set; }
     public string? Name { get; set; }
     public string? Description { get; set; }
-    public float Price { get; set; }
+    public Double Price { get; set; }
 }
